fix failed test
